Add AssessmentState toAssessementService, and refactor saving to API into an implementation method.
