+++ src/core/assessment/AssessmentService.ts
@@ -1,36 +1,9 @@
-import { IApiClient } from '../api/ApiClient';
+import { IAssessmentRemoteClient } from './AssessmentApiClient';
+import { IAssessmentState } from './AssessmentState';
 import { AssessmentInfosRequest } from './dto/AssessmentInfosRequest';
 import { AssessmentResponse } from './dto/AssessmentInfosResponse';
 
-const API_ASSESSMENTS = '/assessments/';
-const ASSESSMENT_VERSION = '1.4.0'; // TODO: Wire this to something automatic.
-
 type AssessmentId = string | null;
-
-export interface IAssessmentRemoteClient {
-  addAssessment(assessment: AssessmentInfosRequest): Promise<AssessmentResponse>;
-  updateAssessment(assessmentId: string, assessment: AssessmentInfosRequest): Promise<AssessmentResponse>;
-}
-
-export class AssessmentApiClient implements IAssessmentRemoteClient {
-  apiClient: IApiClient;
-  constructor(apiClient: IApiClient) {
-    this.apiClient = apiClient;
-  }
-
-  addAssessment(assessment: AssessmentInfosRequest): Promise<AssessmentResponse> {
-    assessment = {
-      ...assessment,
-      version: ASSESSMENT_VERSION,
-    };
-    return this.apiClient.post<AssessmentInfosRequest, AssessmentResponse>(API_ASSESSMENTS, assessment);
-  }
-
-  updateAssessment(assessmentId: string, assessment: AssessmentInfosRequest): Promise<any> {
-    const assessmentUrl = `/assessments/${assessmentId}/`;
-    return this.apiClient.patch<AssessmentInfosRequest, AssessmentResponse>(assessmentUrl, assessment);
-  }
-}
 
 export interface IAssessmentService {
   saveAssessment(assessmentId: AssessmentId, assessment: Partial<AssessmentInfosRequest>): Promise<AssessmentResponse>;
@@ -39,12 +12,17 @@ export interface IAssessmentService {
 
 export default class AssessmentService implements IAssessmentService {
   apiClient: IAssessmentRemoteClient;
+  state: IAssessmentState;
 
-  constructor(apiClient: IAssessmentRemoteClient) {
+  constructor(apiClient: IAssessmentRemoteClient, state: IAssessmentState) {
     this.apiClient = apiClient;
+    this.state = state;
   }
 
-  private async saveToApi(assessmentId: AssessmentId, assessment: Partial<AssessmentInfosRequest>) {
+  private async saveToApi(
+    assessmentId: AssessmentId,
+    assessment: Partial<AssessmentInfosRequest>
+  ): Promise<AssessmentResponse> {
     let response;
     if (assessmentId) {
       response = await this.apiClient.updateAssessment(assessmentId, assessment as AssessmentInfosRequest);
@@ -54,11 +32,15 @@ export default class AssessmentService implements IAssessmentService {
     return response;
   }
 
+  private async saveToState(assessment: Partial<AssessmentInfosRequest>): Promise<boolean> {
+    return this.state.updateAssessment(assessment);
+  }
+
   async saveAssessment(
     assessmentId: AssessmentId,
     assessment: Partial<AssessmentInfosRequest>
   ): Promise<AssessmentResponse> {
-    return this.saveToApi(assessmentId, assessment);
+    return await this.saveToApi(assessmentId, assessment);
   }
 
   async completeAssessment(
